API

Checkpoints

Checkpoints.Checkpoints — Module

Checkpoints

A very minimal module for defining checkpoints or save location in large codebase with the ability to configure how those checkpoints save data externally (similar to how Memento.jl works for logging).

Checkpoints.available — Method

available() -> Vector{String}

Returns a vector of all available (registered) checkpoints.

Checkpoints.checkpoint — Method

checkpoint([prefix], name, data)
checkpoint([prefix], name, data::Pair...; tags...)
checkpoint([prefix], name, data::Dict; tags...)

Defines a data checkpoint with a specified label and values data. By default checkpoints are no-ops and need to be explicitly configured.

checkpoint(session, data; tags...)
checkpoint(handler, name, data::Dict; tags...)

Alternatively, you can also checkpoint with to a session which stages the data to be commited later by commit!(session). Explicitly calling checkpoint on a handler is generally not advised, but is an option.

Checkpoints.config — Method

config(handler::Handler, labels::Vector{String})
config(handler::Handler, prefix::String)
config(labels::Vector{String}, args...; kwargs...)
config(prefix::String, args...; kwargs...)

Configures the specified checkpoints with a Handler. If the first argument is not a Handler then all args and kwargs are passed to a Handler constructor for you.

Checkpoints.register — Function

register([prefix], labels)

Registers a checkpoint that may be configured at a later time.